♻️(GenericTable): add custom renderer option for column
