+++ frontend/src/components/GenericTable/GenericTable.stories.tsx
@@ -2,7 +2,16 @@ import React from 'react';
 
 import { storiesOf } from '@storybook/react';
 import { text, number, date, withKnobs } from '@storybook/addon-knobs';
+import { Cell } from 'components/GenericTable/GenericTable.style';
 import { GenericTable } from './GenericTable';
+
+const defaultCellRenderer = (key: string, value: string, style?: Object): JSX.Element => {
+  return (
+    <Cell key={key} style={style}>
+      {value}
+    </Cell>
+  );
+};
 
 const columnsConfig = [
   {
@@ -10,48 +19,64 @@ const columnsConfig = [
     key: 'id',
     name: 'id',
     columnWidth: 75,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCellRenderer(key, value, style),
   },
   {
     index: 1,
     key: 'body',
     name: 'body',
     columnWidth: 300,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCellRenderer(key, value, style),
   },
   {
     index: 2,
     key: 'filePath',
     name: 'filePath',
     columnWidth: 300,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCellRenderer(key, value, style),
   },
   {
     index: 3,
     key: 'url',
     name: 'url',
     columnWidth: 300,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCellRenderer(key, value, style),
   },
   {
     index: 4,
     key: 'commentor',
     name: 'commentor',
     columnWidth: 200,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCellRenderer(key, value, style),
   },
   {
     index: 5,
     key: 'pullRequestUrl',
     name: 'pullRequestUrl',
     columnWidth: 200,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCellRenderer(key, value, style),
   },
   {
     index: 6,
     key: 'repositoryId',
     name: 'repositoryId',
     columnWidth: 200,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCellRenderer(key, value, style),
   },
   {
     index: 6,
     key: 'creationDate',
     name: 'creationDate',
     columnWidth: 200,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCellRenderer(key, value, style),
   },
 ];
