+++ frontend/src/components/GenericTable/GenericTable.type.ts
@@ -3,7 +3,7 @@ export type ColumnType = {
   key: string;
   name: string;
   columnWidth: number;
-  // renderer: ({ columnIndex, key, rowIndex, style }: RendererInputType) => JSX.Element;
+  renderer: (key: string, value: any, style?: Object) => JSX.Element;
 };
 
 export type RendererInputType = {
