+++ frontend/src/pages/Comments/columnsConfig.tsx
@@ -1,0 +1,80 @@
+import React from 'react';
+import { ColumnType } from 'components/GenericTable/GenericTable.type';
+import { Cell } from 'components/GenericTable/GenericTable.style';
+
+const defaultCommentCellRenderer = (key: string, value: string, style?: Object): JSX.Element => {
+  return (
+    <Cell key={key} style={style}>
+      {value}
+    </Cell>
+  );
+};
+
+export const columnsConfig: ColumnType[] = [
+  {
+    index: 0,
+    key: 'id',
+    name: 'id',
+    columnWidth: 75,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCommentCellRenderer(key, value, style),
+  },
+  {
+    index: 1,
+    key: 'body',
+    name: 'body',
+    columnWidth: 300,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCommentCellRenderer(key, value, style),
+  },
+  {
+    index: 2,
+    key: 'filePath',
+    name: 'filePath',
+    columnWidth: 300,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCommentCellRenderer(key, value, style),
+  },
+  {
+    index: 3,
+    key: 'url',
+    name: 'url',
+    columnWidth: 300,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCommentCellRenderer(key, value, style),
+  },
+  {
+    index: 4,
+    key: 'commentor',
+    name: 'commentor',
+    columnWidth: 200,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCommentCellRenderer(key, value, style),
+  },
+  {
+    index: 5,
+    key: 'pullRequestUrl',
+    name: 'pullRequestUrl',
+    columnWidth: 200,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCommentCellRenderer(key, value, style),
+  },
+  {
+    index: 6,
+    key: 'repositoryId',
+    name: 'repositoryId',
+    columnWidth: 200,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCommentCellRenderer(key, value, style),
+  },
+  {
+    index: 6,
+    key: 'creationDate',
+    name: 'creationDate',
+    columnWidth: 200,
+    renderer: (key: string, value: any, style?: Object): JSX.Element =>
+      defaultCommentCellRenderer(key, value, style),
+  },
+];
+
+export const fixedColumnCount = 2; // this is the reference
